introduce-batch-result-processor-move: Move batch-result processing to SchedulerBatchResultProcessor and retire output_processor mixin
Mechanical cut + paste for the ``introduce-batch-result-processor`` mech
move (final extract from ``SchedulerOutputProcessorMixin``).

Cut the remaining process/collect @staticmethods (after prep) from
``SchedulerOutputProcessorMixin`` and paste them into
``SchedulerBatchResultProcessor`` body in
``scheduler_components/batch_result_processor.py``.

Drop ``@staticmethod`` decorators; simplify
``self: "SchedulerBatchResultProcessor"`` → bare ``self``. Method bodies
byte-identical (all Callable / privacy / mutator substitutions happened in
prep so this step is pure mechanical cut+paste).

Delete the now-empty ``scheduler_output_processor_mixin.py`` file. Drop
the ``SchedulerOutputProcessorMixin`` import + inheritance entry from
``Scheduler``.

All callers updated:
  ``self.<m>(self.batch_result_processor, ...)`` →
  ``self.batch_result_processor.<m>(...)``
(pure prefix transformation).

Refactor chain ID: introduce-batch-result-processor-move
